Include inner class fields in projections
